// src/app/Models/User.ts
export class User {
  id?: number;
  firstName: string | null;
  lastName: string | null;
  phone: string | null;
  email: string | null;
  street: string | null;
  houseNumber: number | null;
  extension: string | null;
  postalCode: string | null;
  city: string | null;
  sizeId: number | null;
  genderId: number | null;
}